[CHORE] Refactor trees
